### 2025-08-21 #8824 @Kevin-Mattheus-Moerman
Note to editors, this is a re-submission of: https://github.com/openjournals/joss-reviews/issues/7654
