### 2025-11-03 #9339 @editorialbot
[REVIEW]: Chemical Recommender System: Replacement Suggestions for Small Molecules
**Submitting author:** @parthivn25 (<a href="http://orcid.org/0000-0002-2533-7895">Parthiv Nair</a>)
**Repository:** https://github.com/sandialabs/chemical-recommender-system
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.2
