### 2025-11-22 sandialabs/chemical-recommender-system#1 @andrewtarzia
Minor comment on JOSS paper
While the paper outlines clear the technical improvements and why CRS is impressive, it does not offer a "non-specialist" audience "how to use" statement, or link to the stated web application (unless that is the locally run docker image?)

Review part of https://github.com/openjournals/joss-reviews/issues/9339
