### 2025-11-22 sandialabs/chemical-recommender-system#2 @andrewtarzia
Benchmarks missing from JOSS paper
As requested in the checklist, the "state of the field" is necessary, and only touched on by stating the presence of other recommender systems. Efficiency is mentioned:

> Our experiments demonstrate that this infrastructure enables searches over the whole dataset that complete in minutes rather than hours on standard hardware.

But the details of what was being compared are missing. I do not think these are critical issues, and I appreciate that the improvements in CRS are in the openness and extensibility. Is there a simple way to show the accuracy of the CRS outputs? Figure 3 may do this, but is not clear enough to interpret the list of molecules. Perhaps just showing what the molecules in that list are underneath Figure 3 would be sufficient to provide confidence?

Review part of https://github.com/openjournals/joss-reviews/issues/9339
