### 2025-11-22 sandialabs/chemical-recommender-system#3 @andrewtarzia
Testing use of CRS through docker
- I recommend writing instructions and docs as lists of instructions rather than full paragraphs. These are easier to follow, and helps when users are not familiar with docker.
- Significant download time the first time I ran `docker compose up -d` - it is a 17GB download. I do not use docker really but that seems like a significant amount of data to expect users put on their systems? Especially without warnings in the docs.
- I opened `localhost:5005:` in my browser, the web app looks very good.
- Testing with `1923`: gave an error: `PUGREST Timeout`
- Testing smiles `c1ccccc1`: This ran (I had to rerun after another timeout; it took a few minutes on my standard laptop). PDF report and CSV data not to `app/output` in container.
- Testing with the command line interface was smooth (this saved to `output`). Although it seemed that connectivity was a constant issue for me (I saw `ERROR - Search attempt 1 failed: <MilvusException: (code=101, message=failed to search: collection not loaded[collection=459572460743118123])>` many times), which was handled nicely in the CLI. The output extracted from the container though was for a different query than the one I coped from the README (it was for one I had submitted earlier). I reran the test and the output now matched the expected query.
- I do not have the expertise to review the docker commands.

Review part of https://github.com/openjournals/joss-reviews/issues/9339
