### 2025-11-22 sandialabs/chemical-recommender-system#4 @andrewtarzia
Tests are not automated online for evaluation
As stated in the title. I do not know if this is necessary to pass JOSS review, but there are not automated tests that I can see.

Review part of https://github.com/openjournals/joss-reviews/issues/9339
