### 2025-12-14 sandialabs/chemical-recommender-system#3 @kevingreenman
- I second the comment to write instructions as lists instead of paragraphs, as well as the suggestion to add a mention to the README of the expected total size of the downloads.
- I was not able to open `localhost:5005` in my browser (it did not open automatically, and nothing loaded when I tried navigating there manually). I've only used docker a couple times before so it's possible that I'm doing something wrong, but I have Docker Desktop installed (as well as Rosetta) on my M1 Macbook and tried to follow the instructions in the README. First I tried without an .env file, then I created one to try changing the ports to see if that would fix anything, but it didn't. Docker Desktop shows that the container is running, but nothing is loading for me at `localhost:5005` .
