### 2025-12-14 sandialabs/chemical-recommender-system#5 @kevingreenman
Confusing ordering of README sections
It seems that the order of sections in the README might have gotten mixed up. For example:

- There is a link to "Table of Contents" as the 3rd bullet in the Table of Contents, even though the TOC is first

- The "Environment Variables" section seems misplaced, in between the Input File Format and the descriptions of the parameters in the input file format. Was "Environment Variables" intended to be further up, closer to "Instructions to Run" which references it?

Review part of [openjournals/joss-reviews#9339](https://github.com/openjournals/joss-reviews/issues/9339)
