### 2025-12-14 #9339 @kevingreenman
@parthivn25 regarding the reviewer checklist points about references, state of the field, and statement of need, I was wondering if you could clarify exactly what you mean by "recommender system". You cite DeepChem as an example, but I'm not sure if this is appropriate as I think this would imply that any chemical machine learning model / package / architecture suite counts as a "chemical recommender system".
