### 2026-01-10 34f2d94
readme clarity fix (#5)
