### 2026-01-10 8fd02ae
general explanation (#1)
