+++ paper.md
@@ -34,7 +34,7 @@ The Chemical Recommender System (CRS) is an open-source, high-performance toolki
 
 # Statement of need
 
-Small molecules frequently become 'at risk' or unavailable due to supply, legislative, or technical issues (e.g., the recent discontinuation of PFAS by 3M). In such cases, identifying replacements quickly becomes essential. The CRS has been created to provide a first step in the down-selection of possible replacements for a given target molecule.
+Small molecules frequently become 'at risk' or unavailable due to supply, legislative, or technical issues (e.g., the recent discontinuation of PFAS by 3M). In such cases, identifying replacements quickly becomes essential. The CRS has been created to provide a first step in the down-selection of possible replacements for a given target molecule. To use the system, users input a target molecule using SMILES notation or a PubChem identifier through an intuitive web interface, which then performs rapid similarity searches and ranks potential replacements based on multiple criteria.
 
 Chemical recommender systems have been explored in various forms, such as Recommender Systems for Organic Compounds [@Hayashi:2022], which used machine learning techniques for classification, and DeepChem [@Ramsundar:2017], which applies deep learning models to predict solubility and toxicity in drug discovery. However, existing solutions face significant limitations: commercial platforms like PubChem and SigmAldrich provide black-box similarity searches without technical transparency, configurability, or extensibility; academic tools typically operate on limited datasets and lack the infrastructure to handle full-scale chemical databases; and most systems do not offer seamless integration of custom predictive models. The CRS addresses these gaps by providing a completely open-source platform with full configurability of search parameters, filtering logic, and scoring functions. The system's plug-and-play Docker architecture enables researchers to integrate their own models while leveraging the complete PubChem-scale search infrastructure, creating a unique combination of scale, transparency, and extensibility not available in existing solutions.
 
