### 2026-01-12 #9339 @parthivn25
> [@parthivn25](https://github.com/parthivn25) regarding the reviewer checklist points about references, state of the field, and statement of need, I was wondering if you could clarify exactly what you mean by "recommender system". You cite DeepChem as an example, but I'm not sure if this is appropriate as I think this would imply that any chemical machine learning model / package / architecture suite counts as a "chemical recommender system".

Thanks for pointing out this ambiguity. We agree that DeepChem is not itself a recommender system but rather a machine-learning/QSAR toolkit whose models can be used within such a pipeline, and its previous inclusion as an example was misleading. We have [updated](https://github.com/sandialabs/chemical-recommender-system/commit/e5b3985a08623838faac423c3accddc2bb93ff74) the text accordingly by separating these categories and revising the related-work discussion to reflect this distinction.
