### 2026-01-13 e4a3743
instructions to list rather than paragraph (#3)
