### 2026-01-16 sandialabs/chemical-recommender-system#4 @andrewtarzia
Sorry. I mean a file to create an environment to run the tests (i.e., installing the dependencies).

Did I miss instructions to run the tests locally?
