### 2026-01-16 sandialabs/chemical-recommender-system#4 @andrewtarzia
So it's not possible to run the tests outside the docker container without a 17gb installation?
