### 2026-01-16 sandialabs/chemical-recommender-system#3 @parthivn25
The saving of app/output is expected behavior. When the query is ran through the UI from a browser, the output data is shown on the UI. There are provided links on the UI to download the data from there. The use of app/output is only for queries in which users run the CRS through the terminal by entering the docker container manually. In this situation it is saved to output so a user can copy it over into their local machine for viewing/use.
