### 2026-01-16 63a734d
docker image size note (#3)
