### 2026-01-16 sandialabs/chemical-recommender-system#4 @parthivn25
No, the tests rely on the entire flow of the internal ML models as well as the vector database of all the pubchem molecules. These two entities make up a majority of the installation, which is what we are packaging together and providing for users. Any test that didn't use this data wouldn't truly demonstrate the functionality of the CRS.
