### 2026-01-16 9211e71
tests info in readme (#4)
