### 2026-01-20 443c9b4
tests readme update to use w/o Docker (#4)
