### 2026-01-27 #9339 @espottesmith
@andrewtarzia, @kevingreenman could you please take another look at the paper/code and see if it addresses your issues?
