### 2026-02-11 #9339 @kevingreenman
@espottesmith sorry for the further delay. I tried the installation again and made it further this time, but am still running into some more issues with the tests and functionality. I'm reporting these in https://github.com/sandialabs/chemical-recommender-system/issues/3
