### 2026-03-09 #9339 @parthivn25
Hi, I had made updates in the CRS issue channel and am waiting on a response - https://github.com/sandialabs/chemical-recommender-system/issues/3#issuecomment-3976014612
