### 2026-03-13 sandialabs/chemical-recommender-system#3 @parthivn25
Glad to hear it's working! Thanks for that catch. It looks like the golden test is trying to match not only the data, but also the time/date between runs, which is obviously incorrect. I'll fix that.

I'll also add a note about the proxies and the fix we used for you into the readme.
