### 2026-03-14 sandialabs/chemical-recommender-system#3 @kevingreenman
The `NO_PROXY` change also seems to fix most things on my M1 Mac. The web app and CLI work as expected now. The unit tests also succeed. But the golden master test gives this:
```
FAIL: test_exact_cid_6517_batch_reproduction (tests.test_golden_master.TestGoldenMaster)
Test that reproduces your exact CID 6517 search using BatchRun
Traceback (most recent call last):
  File "/app/tests/test_golden_master.py", line 126, in test_exact_cid_6517_batch_reproduction
    self.assertAlmostEqual(float(actual_row.Overall), float(golden_row.Overall),
AssertionError: 1.2150027681836784 != 1.186905477157309 within 5 places (0.028097291026369264 difference) : CID 524710: Overall score mismatch

Ran 1 test in 375.527s

FAILED (failures=1)
```
