### 2026-03-14 #9339 @kevingreenman
We're very close now, I think the installation issues are pretty much all resolved. Just waiting on updates for a couple small error fixes related to the tests, and updates to the installation instructions and/or file to reflect the fix that worked for me. Then I can check off the last box on my checklist.
