### 2026-03-24 64c9cbc
readme proxy troubleshoot note (#3)
