### 2026-03-24 6cbcbfe
golden test simplification/fix (#3)
