+++ tests/test_golden_master.py
@@ -1,9 +1,7 @@
 # Golden master tests - compare against known good outputs
 import unittest
-import json
 import os
 import sys
-from unittest.mock import patch, Mock
 import pandas as pd
 
 # Add src to path
@@ -15,57 +13,9 @@ class TestGoldenMaster(unittest.TestCase):
     """
     Golden Master tests - compare outputs against known good results.
     
-    These tests use pre-recorded 'golden' outputs to ensure the system
-    produces consistent results. When you make changes that intentionally
-    change outputs, you'll need to update the golden files.
+    This test validates generated batch output against the committed
+    tests/golden.csv baseline.
     """
-    
-    def setUp(self):
-        """Set up paths for golden master files"""
-        self.test_data_dir = os.path.join(os.path.dirname(__file__), 'golden_data')
-        os.makedirs(self.test_data_dir, exist_ok=True)
-    
-    def save_golden_master(self, test_name, data):
-        """Save golden master data to file"""
-        filepath = os.path.join(self.test_data_dir, f"{test_name}.json")
-        with open(filepath, 'w') as f:
-            json.dump(data, f, indent=2, default=str)
-    
-    def load_golden_master(self, test_name):
-        """Load golden master data from file"""
-        filepath = os.path.join(self.test_data_dir, f"{test_name}.json")
-        if not os.path.exists(filepath):
-            return None
-        with open(filepath, 'r') as f:
-            return json.load(f)
-    
-    def assert_golden_master(self, test_name, actual_data, tolerance=1e-6):
-        """Compare actual data against golden master"""
-        expected_data = self.load_golden_master(test_name)
-        
-        if expected_data is None:
-            # First time running - save as golden master
-            self.save_golden_master(test_name, actual_data)
-            print(f"Saved new golden master: {test_name}")
-            return
-        
-        # Compare the data structures
-        self.compare_data_structures(expected_data, actual_data, tolerance)
-    
-    def compare_data_structures(self, expected, actual, tolerance):
-        """Recursively compare data structures with tolerance for floats"""
-        if isinstance(expected, dict) and isinstance(actual, dict):
-            self.assertEqual(set(expected.keys()), set(actual.keys()))
-            for key in expected.keys():
-                self.compare_data_structures(expected[key], actual[key], tolerance)
-        elif isinstance(expected, list) and isinstance(actual, list):
-            self.assertEqual(len(expected), len(actual))
-            for exp_item, act_item in zip(expected, actual):
-                self.compare_data_structures(exp_item, act_item, tolerance)
-        elif isinstance(expected, float) and isinstance(actual, float):
-            self.assertAlmostEqual(expected, actual, delta=tolerance)
-        else:
-            self.assertEqual(expected, actual)
 
     def test_exact_cid_6517_batch_reproduction(self):
         """
@@ -154,40 +104,6 @@ class TestGoldenMaster(unittest.TestCase):
             self.assertAlmostEqual(float(top_result['Fingerprint']), float(golden_top['Fingerprint']), places=4,
                                  msg="Top result fingerprint score should match golden data")
             
-            # Store comprehensive summary for regression testing
-            test_summary = {
-                'query_cid': 6517,
-                'batch_parameters': {
-                    'finnum': 30,
-                    'tarray': [True, True, False, False, False],
-                    'incEle': False,
-                    'include_specific_elements': ['Si'],
-                    'disallow_isotopes': 1,
-                    'substructure': 'CCO',
-                    'substructure_count': 1
-                },
-                'num_results': len(actual_ranked_results),
-                'all_result_cids': list(actual_ranked_results['CID'].astype(int)),
[…]
