### 2026-03-25 sandialabs/chemical-recommender-system#3 @kevingreenman
It looks like your recent commits have addressed both of my remaining concerns, thanks! I would close the issue, but it looks like I don't have permission to do so, so feel free to do that yourself @parthivn25
