### 2026-03-25 #9339 @kevingreenman
Hi @espottesmith, it looks like @parthivn25 made some commits yesterday that addressed my last two remaining concerns we were discussing on the thread at https://github.com/sandialabs/chemical-recommender-system/issues/3. So I checked off my last checkbox.
