### 2026-03-25 #9339 @parthivn25
Hi espottesmith,

1. I've checked the author list.
2. I made a release of the software under v1.2
3. I archived the release on Zenodo with DOI: https://doi.org/10.5281/zenodo.19227441.
4. The archive authors/orcids match that of the paper exactly.
5. The license for the CRS and archive are both BSD-3.
