### 2026-03-26 #9339 @espottesmith
Some small comments on the manuscript:

- "the recent discontinuation of PFAS by 3M" should have a reference
- Readers might appreciate references to the various tools you use, as well, e.g., etcd, RDkit, and/or PubChem

Otherwise, the paper looks great. It's quite well written.
