### 2026-03-26 #9339 @parthivn25
Thank you for the helpful suggestions. I added a reference for the 3M/PFAS example in the Statement of Need, and I added in-text citations for PubChem and RDKit where they are introduced in the manuscript. The manuscript already cited the other core technologies and methods used by the CRS, including Milvus, OPERA, fingerprint similarity methods, and synthetic accessibility scoring.
