### 2026-03-26 #9339 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
