### 2026-03-30 #9339 @Kevin-Mattheus-Moerman
@parthivn25 As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:

👉 It looks like this DOI may be suitable instead of the ArXiv citation you use: https://doi.org/10.1002/adma.201604837
- [x] Copyedit pass/checked for typos
