### 2026-04-17 80b7d38
Add DOI badge to README
Added a DOI badge to the README for citation purposes.
